Add files via upload
Q1. Create a function which will take a list as an argument and return the product of all the numbers
after creating a flat list.

Use the below-given list as an argument for your function.


list1 = [1,2,3,4, [44,55,66, True], False, (34,56,78,89,34), {1,2,3,3,2,1}, {1:34, "key2": [55, 67, 78, 89], 4: (45,
22, 61, 34)}, [56, 'data science'], 'Machine Learning']


Note: you must extract numeric keys and values of the dictionary also.
Q2. Write a python program for encrypting a message sent to you by your friend. The logic of encryption
should be such that, for a the output should be z. For b, the output should be y. For c, the output should
be x respectively. Also, the whitespace should be replaced with a dollar sign. Keep the punctuation
marks unchanged.


Input Sentence: I want to become a Data Scientist.


Encrypt the above input sentence using the program you just created.